pyproject.toml
[project]
name = "ninjaone-activity-downloader"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "httpx==0.27.0",
    "python-dotenv==1.0.1",
    "rich>=14.3.3",
]
